Move loading screen to sub-namespace
