+++ AssetHelper/Plugin/Tasks/BaseStartupTask.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Silksong.AssetHelper.Plugin.LoadingPage;
+using System.Collections;
 
 namespace Silksong.AssetHelper.Plugin.Tasks;
