+++ AssetHelper/Plugin/Tasks/NonSceneCatalog.cs
@@ -1,5 +1,6 @@
 ﻿using Silksong.AssetHelper.Core;
 using Silksong.AssetHelper.Internal;
+using Silksong.AssetHelper.Plugin.LoadingPage;
 using System;
 using System.Collections;
 using System.Collections.Generic;
